Get project compiling (but not linking... need to build middlewares) for OsX

+++ Source/Core/SystemLib/MemoryAllocation.h
@@ -4,6 +4,8 @@
 // Joe Schutte
 //==============================================================================
 
+#include <SystemLib/BasicTypes.h>
+
 namespace Selas
 {
     #define PlacementNew_(Type_, Var_)                     new(Var_) Type_()
@@ -24,9 +26,9 @@ namespace Selas
     #define FreeAligned_(Var_)                             Selas::SelasAlignedFree(Var_)
     #define SafeFreeAligned_(Var_)                         if(Var_) { Selas::SelasAlignedFree(Var_); Var_ = nullptr; }
 
-    extern void* SelasAlignedMalloc(size_t size, size_t alignment, const char* name, const char* file, int line);
-    extern void* SelasMalloc(size_t size, const char* name, const char* file, int line);
-    extern void* SelasRealloc(void* address, size_t size, const char* name, const char* file, int line);
+    extern void* SelasAlignedMalloc(uint32 size, uint32 alignment, const char* name, const char* file, int line);
+    extern void* SelasMalloc(uint32 size, const char* name, const char* file, int line);
+    extern void* SelasRealloc(void* address, uint32 size, const char* name, const char* file, int line);
     extern void  SelasAlignedFree(void* address);
     extern void  SelasFree(void* address);
